유저추가 발걸음추가

+++ src/main/java/healthcare/healthcare_spring/domain/CalculateRepository.java
@@ -2,5 +2,5 @@ package healthcare.healthcare_spring.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CalculateRepository extends JpaRepository<Calculate, Long> {
+public interface CalculateRepository extends JpaRepository<Calculate, Integer> {
 }
